Trait sc_proof_of_time::PotConsensusState

pub trait PotConsensusState: Send + Sync {
    // Required methods
    fn get_block_proofs<'life0, 'life1, 'async_trait>(
        &'life0 self,
        block_number: BlockNumber,
        current_slot: SlotNumber,
        parent_pre_digest: &'life1 PotPreDigest,
        wait_time: Option<Duration>
    ) -> Pin<Box<dyn Future<Output = Result<NonEmptyVec<PotProof>, PotGetBlockProofsError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn verify_block_proofs(
        &self,
        block_number: BlockNumber,
        slot_number: SlotNumber,
        pre_digest: &PotPreDigest,
        parent_slot_number: SlotNumber,
        parent_pre_digest: &PotPreDigest
    ) -> Result<(), PotVerifyBlockProofsError>;
}

Interface to consensus.

Required Methods

fn get_block_proofs<'life0, 'life1, 'async_trait>( &'life0 self, block_number: BlockNumber, current_slot: SlotNumber, parent_pre_digest: &'life1 PotPreDigest, wait_time: Option<Duration> ) -> Pin<Box<dyn Future<Output = Result<NonEmptyVec<PotProof>, PotGetBlockProofsError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Called by consensus when trying to claim the slot. Returns the proofs in the slot range [start_slot, current_slot - global_randomness_reveal_lag_slots]. If the proofs are unavailable, retries upto wait_time(if specified).

fn verify_block_proofs( &self, block_number: BlockNumber, slot_number: SlotNumber, pre_digest: &PotPreDigest, parent_slot_number: SlotNumber, parent_pre_digest: &PotPreDigest ) -> Result<(), PotVerifyBlockProofsError>

Called during block import validation. Verifies the sequence of proofs in the block being validated.