sc_proof_of_time

Trait PotSlotWorker

pub trait PotSlotWorker<Block>
where
    Block: BlockT,
{
    // Required method
    fn on_proof(&mut self, slot: Slot, checkpoints: PotCheckpoints);
}

Required Methods

fn on_proof(&mut self, slot: Slot, checkpoints: PotCheckpoints)

Called when new proof of time is available for slot.

NOTE: Can be called more than once in case of reorgs to override old slots.